+++ src/js/form.js
@@ -438,7 +438,7 @@ Form.prototype.setAllVals = function( $group, groupIndex ) {
                 const index = that.model.node( name ).getElements().indexOf( element );
                 const control = that.input.find( name, index );
                 if ( control ) {
-                    that.input.setVal( control, value );
+                    that.input.setVal( control, value, null );
                     if ( that.input.getXmlType( control ) === 'binary' && value.startsWith( 'jr://' ) && element.getAttribute( 'src' ) ) {
                         control.setAttribute( 'data-loaded-url', element.getAttribute( 'src' ) );
                     }