+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Changed
+- No longer triggering an inputupdate event when setting default values in view.
+
 [5.11.0] - 2019-12-31
 ---------------------
 ##### Added